requirements.txt
django-extensions
django-fortune
